% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zAMPExplorer_app.R
\name{zAMPExplorer}
\alias{zAMPExplorer}
\title{Launch the zAMPExplorer Shiny App}
\usage{
zAMPExplorer()
}
\description{
This function launches the Shiny app for zAMPExplorer. It assumes that the app files
(ui.R and server.R) are located in the \code{inst/shiny} directory.
}
